+++ packages/core/lib/index.js
@@ -2,10 +2,12 @@ const createConfig = require('./createConfig')
 const group = require('./group')
 const env = require('./env')
 const match = require('./match')
+const when = require('./when')
 
 module.exports = {
   createConfig,
   group,
   env,
-  match
+  match,
+  when
 }